+++ gridy/Sources/Presentations/PlanBoard/PlanItem.swift
@@ -1,5 +1,5 @@
 //
-//  TimelineLayoutReducer.swift
+//  PlanItem.swift
 //  gridy
 //
 //  Created by 제나 on 2023/09/26.
